requirements.txt
matplotlib==1.5.3
numpy==1.13.1
scipy==0.18.1
tensorflow_gpu==1.1.0
Pillow==4.2.1
tensorflow==1.3.0
